Don't crash on empty pages (#32)
* Don't crash on empty pages

I've got a page on my site that has only an iframe. Before this
commit, wp2hugo would crash when it reached that page but with this
change it emits a warning and keeps going, which in my case allows it
to finish processing the site.

* Don't warn about empty footnotes

wp2hugo handled empty footnotes but it emitted a warning about "Error
unmarshalling footnotes". Now it skips them entirely.

* Re-add empty footnote warning