fix filenames for media with query values
wordpress.com's CMS automatically adds query values for images in a few
cases.

when wp2hugo downloads media of  URLs with query values, e.g.
`foo.png?q=1` migrated pages fail to load this media.

the media file is stored with the filename that includes the query
values (foo.png?q=1) but the markdown only refers to the path (foo.png).

this patch removes the query values from the filenames on disk.