fix: fix a panic in Gallery shortcode
If the WordPress `gallery` shortcode is missing `ids` then persist the
original shortcode in the generated Markdown code.

Ref: https://github.com/ashishb/wp2hugo/issues/64
From the original comment

```
I suspect this is caused by the importer assuming that a gallery shortcode will have a list of IDs, however, that's not always the case. My export contains gallery shortcodes in the following format (the link="file" bit I don't think is relevant, but I've included it for completeness).

[gallery link="file"]

That will include all media that is attached to the post by default on WordPress.
```