Support WPML translations using parameter URL scheme
For reference : https://wpml.org/documentation/getting-started-guide/language-setup/language-url-options/#language-name-added-as-a-parameter

This slightly changes the Markdown files naming scheme, trying first to re-use the slug of the WordPress page (last member of the URL path). If no slug is found, this defaults to the current logic, using normalized title. Then, we try to extract the ?lang=XX or &lang=XX parameter if any.

The best case scenario is this :

- WP input : `https://website.com/some-page/` -> Hugo file `content/pages/some-page.md`,
- WP input `https://website.com/some-page/?lang=fr` -> Hugo file `content/pages/some-page.fr.md`

This automatically follows Hugo translating scheme. However, if the page slug was also translated through WPML, this doesn't work as-is anymore.